+++ puffin_egui/CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to the egui crate will be documented in this file.
 * Show scrollbar for history of recent frames.
 * Show date-time of when a frame was recorded.
 * Show compressed size of selected frame.
+* Fix occasional flickering when viewing merged scopes.
 
 
 ## 0.10.0 - 2021-10-29